Allow attachments in NavigatorView._get_page_payload
The previous behaviour for send_page was to set attachment to None
(and to error out) when _get_page_payload returned "attachments" if
overridden by a user.
This behaviour has been changed to only do so when no attachments are
returned by the _get_page_payload method. This change allows views
with attachments by overriding only _get_page_payload.